+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Update :meth:`api.utils.estimate_phase_beagle` method to warn when statistical phasing is skipped.
 * Upgrade Beagle version from v5.2 (beagle.28Jun21.220.jar) to v5.4 (beagle.22Jul22.46e.jar) due to a bug in v5.2.
 * Update :meth:`api.utils.estimate_phase_beagle` method to filter out variants with improper allele ('I', 'D', 'N', '.'). Note that this issue is specific to chip data.
+* Update :meth:`api.utils.import_variants` method to handle input VCF with duplicate variants. Basically, it will warn the user about it and and only keep the first record. This issue seems to occur frequently with chip data.
 
 0.17.0 (2022-07-12)
 -------------------